Move encryption/decryption stuff to common.
