+++ plugins/common/encryption/aes.go
@@ -1,4 +1,4 @@
-package http
+package encryption
 
 import (
 	"crypto/aes"
