+++ plugins/common/encryption/encryption.go
@@ -1,0 +1,6 @@
+package encryption
+
+type Decryptor interface {
+	Init() error
+	Decrypt(data []byte) ([]byte, error)
+}
