+++ plugins/common/encryption/padding.go
@@ -1,4 +1,4 @@
-package http
+package encryption
 
 func PKCS7Trimming(in []byte) []byte {
 	// 'count' number of bytes where padded to the end of the clear-text
